chore:refactor code
